storage-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1

# S3 / AWS
boto3==1.34.58

# MongoDB (choose one, motor is async)
# pymongo==4.6.2
motor==3.3.2 # For async MongoDB operations with FastAPI

# File handling
python-multipart==0.0.9 # For FastAPI file uploads
python-magic==0.4.27 # For MIME type detection from file content (more robust)

# Image Processing (optional, if advanced processing is needed beyond basic uploads)
Pillow==10.2.0 # For image manipulation, thumbnails

# Scheduled tasks (if not using a central Celery)
# apscheduler==3.10.4

# Logging, Metrics, Tracing
structlog==24.1.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-boto3sqs ==0.4.0 # If using SQS with Boto3
opentelemetry-instrumentation-pymongo==0.44b0 # If using pymongo
opentelemetry-instrumentation-motor==0.2.0 # If using motor
starlette-prometheus==0.10.0

# Linters and formatters (dev)
# flake8
# black
# isort
# mypy

# For JWT handling
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 # For password hashing, if needed (likely in auth-service more)